quicksort2

+++ ParallelMergeSort/src/prog/Utilities.java
@@ -38,7 +38,7 @@ public class Utilities {
 
     public static void printArray(float[] numbers) {
         for (int i = 0; i < numbers.length; i++) {
-            System.out.print(numbers[i] + ", ");
+            System.out.print((int)numbers[i] + ", ");
         }
     }
